+++ Fly to You/Fly to You/Presentation/Component/Cell/PlaneCell.swift
@@ -61,12 +61,6 @@ struct PlaneCell: View{
     }
 }
 
-enum PlaneCellRoute{
-    case receive
-    case send
-    case map
-}
-
 //#Preview {
 //    PlaneCell()
 //}
