+++ Fly to You/Fly to You/Presentation/Component/SearchBar.swift
@@ -36,7 +36,3 @@ struct SearchBar: View {
     }
 }
 
-enum searchBarRoute{
-    case searchNickname
-    case searchTopic
-}
