+++ Fly to You/Fly to You/Presentation/Scene/DepartureLogScene/View/DepartureLogView.swift
@@ -84,7 +84,3 @@ class DepatureLogViewModelWrapper: ObservableObject {
 #Preview {
     DepartureLogView()
 }
-
-enum DepartureLogRoute {
-    case departureLogInfo
-}
