+++ Fly to You/Fly to You/Presentation/Scene/FlightMapScene/View/FlightMapView.swift
@@ -79,7 +79,7 @@ struct FlightMapView: View{
             viewModelWrapper.viewModel.removeFlightsListener()
         }
         .sheet(isPresented: $showReportModal) {
-            ReportSheetView()
+            ReportSheetView(letter: selectedRoute)
                 .presentationDragIndicator(.visible)
         }
     }
