+++ Fly to You/Fly to You/Presentation/Scene/SendScene/View/SendLetterView.swift
@@ -99,8 +99,3 @@ struct SendLetterView: View{
     }
 }
 
-enum SendLetterRoute {
-    case start
-    case relay
-}
-
